+++ LED.ino
@@ -1,9 +1,9 @@
 #define red 9
 #define green 10
 #define blue 11
-int rgbBrightness[3];
-int fadeAmount = 1;
-byte maxBrightness = 50;
+uint8_t rgbBrightness[3];
+uint8_t fadeAmount = 1;
+uint8_t  maxBrightness = 50;
 
 unsigned long previousUpdateMillis;
